+++ apps/backend/src/modules/users/user.controller.ts
@@ -65,6 +65,46 @@ class UserController extends BaseController {
 				body: userPatchValidationSchema,
 			},
 		});
+
+		this.addRoute({
+			handler: (options) =>
+				this.delete(
+					options as APIHandlerOptions<{
+						params: { id: string };
+					}>,
+				),
+			method: "DELETE",
+			path: UsersApiPath.$ID,
+		});
+	}
+
+	/**
+	 * @swagger
+	 * /users/{id}:
+	 *    delete:
+	 *      tags:
+	 *        - Users
+	 *      description: Deletes a user
+	 *      parameters:
+	 *        - in: path
+	 *          name: id
+	 *          description: ID of the user to delete
+	 *          required: true
+	 *          schema:
+	 *            type: string
+	 *      responses:
+	 *        204:
+	 *          description: User deleted successfully
+	 */
+	private async delete(
+		options: APIHandlerOptions<{
+			params: { id: string };
+		}>,
+	): Promise<APIHandlerResponse> {
+		return {
+			payload: await this.userService.delete(Number(options.params.id)),
+			status: HTTPCode.OK,
+		};
 	}
 
 	/**