+++ apps/backend/src/modules/users/user.repository.ts
@@ -31,14 +31,22 @@ class UserRepository implements Repository {
 		return UserEntity.initialize(user);
 	}
 
-	public delete(): ReturnType<Repository["delete"]> {
-		return Promise.resolve(true);
+	public async delete(id: number): Promise<boolean> {
+		const deletedRowsCount = await this.userModel
+			.query()
+			.patch({ deletedAt: new Date().toISOString() })
+			.where({ id })
+			.whereNull("deletedAt")
+			.execute();
+
+		return Boolean(deletedRowsCount);
 	}
 
 	public async find(id: number): Promise<null | UserEntity> {
 		const user = await this.userModel
 			.query()
 			.findById(id)
+			.whereNull("deletedAt")
 			.returning("*")
 			.execute();
 
@@ -51,6 +59,7 @@ class UserRepository implements Repository {
 	}: PaginationQueryParameters): Promise<PaginationResponseDto<UserEntity>> {
 		const { results, total } = await this.userModel
 			.query()
+			.whereNull("deletedAt")
 			.page(page, pageSize);
 
 		return {
@@ -59,8 +68,17 @@ class UserRepository implements Repository {
 		};
 	}
 
-	public async findByEmail(email: string): Promise<null | UserEntity> {
-		const user = await this.userModel.query().findOne({ email });
+	public async findByEmail(
+		email: string,
+		hasDeleted = false,
+	): Promise<null | UserEntity> {
+		const query = this.userModel.query().findOne({ email });
+
+		if (!hasDeleted) {
+			query.whereNull("deletedAt");
+		}
+
+		const user = await query.execute();
 
 		return user ? UserEntity.initialize(user) : null;
 	}