+++ apps/backend/src/modules/users/user.service.ts
@@ -31,7 +31,7 @@ class UserService implements Service {
 		payload: UserSignUpRequestDto,
 	): Promise<UserAuthResponseDto> {
 		const { email, name, password } = payload;
-		const existingUser = await this.userRepository.findByEmail(email);
+		const existingUser = await this.userRepository.findByEmail(email, true);
 
 		if (existingUser) {
 			throw new UserError({
@@ -55,8 +55,17 @@ class UserService implements Service {
 		return item.toObject();
 	}
 
-	public delete(): ReturnType<Service["delete"]> {
-		return Promise.resolve(true);
+	public async delete(id: number): Promise<boolean> {
+		const isDeleted = await this.userRepository.delete(id);
+
+		if (!isDeleted) {
+			throw new UserError({
+				message: ExceptionMessage.USER_NOT_FOUND,
+				status: HTTPCode.NOT_FOUND,
+			});
+		}
+
+		return isDeleted;
 	}
 
 	public async find(id: number): Promise<UserAuthResponseDto> {
@@ -91,7 +100,7 @@ class UserService implements Service {
 
 		if (!item) {
 			throw new UserError({
-				message: ExceptionMessage.USER_NOT_FOUND,
+				message: ExceptionMessage.INVALID_CREDENTIALS,
 				status: HTTPCode.NOT_FOUND,
 			});
 		}