+++ apps/frontend/src/libs/modules/api/base-http-api.ts
@@ -75,7 +75,7 @@ class BaseHTTPApi implements HTTPApi {
 		if (hasAuth) {
 			const token = await this.storage.get<string>(StorageKey.TOKEN);
 
-			headers.append(HTTPHeader.AUTHORIZATION, `Bearer ${token ?? ""}`);
+			headers.append(HTTPHeader.AUTHORIZATION, `Bearer ${token as string}`);
 		}
 
 		return headers;